backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
yt-dlp
ffmpeg-python==0.2.0
openai==1.10.0
reportlab==4.0.9
arabic-reshaper==3.0.0
python-bidi==0.4.2
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
